refactor(frontend): Add Viewer component and related files
