+++ frontend/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Home from './components/Home';
-import Viewer from './components/Viewer';
+import Viewer from './components/Viewer/Viewer';
 
 function App() {
   return (
